+++ hercules-ci-agent/CHANGELOG.md
@@ -24,6 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - File lookup order has changed, to support flakes. `ci.nix` or `nix/ci.nix` still take top priority, followed by `flake.nix`, followed by `default.nix`.
 
+### Fixed
+
+ - When the root of a `ci.nix` is a list, an error message is returned.
 
 ## [0.8.3]