+++ hercules-ci-cnix-expr/CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
  - `ToValue` and `ToRawValue` classes for converting Haskell values to Nix values.
 
+ - `Hercules.CNix.Expr.Schema` module for a typed interface between code in the Nix language and Haskell. This also includes provenance tracking, improving error messages while reducing error handling noise in the Haskell code.
+
 ## 0.2.0.0 - 2021-06-22
 
 ### Added